adjusted profiles routes

+++ src/routes/profiles/index.ts
@@ -39,6 +39,18 @@ const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
       },
     },
     async function (request, reply): Promise<ProfileEntity> {
+      const userProfile = await this.db.profiles.findOne({ key: 'userId', equals: request.body.userId })
+
+      if (request.body.memberTypeId !== 'basic' && request.body.memberTypeId !== 'business') {
+        reply.code(400)
+        throw new Error('Wrong memberTypeId supplied')
+      }
+
+      if (userProfile) {
+        reply.code(400)
+        throw new Error('User already has a profile')
+      }
+
       return this.db.profiles.create(request.body)
     }
   );
@@ -51,6 +63,13 @@ const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
       },
     },
     async function (request, reply): Promise<ProfileEntity> {
+      const profile = await this.db.profiles.findOne({ key: 'id', equals: request.params.id })
+
+      if (!profile) {
+        reply.code(400)
+        throw new Error('Profile was not found')
+      }
+
       return this.db.profiles.delete(request.params.id)
     }
   );
@@ -64,6 +83,13 @@ const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
       },
     },
     async function (request, reply): Promise<ProfileEntity> {
+      const profile = await this.db.profiles.findOne({ key: 'id', equals: request.params.id })
+
+      if (!profile) {
+        reply.code(400)
+        throw new Error('Profile was not found')
+      }
+
       return this.db.profiles.change(request.params.id, request.body)
     }
   );
